forgive .0 in output

+++ src/Main.java
@@ -20,7 +20,7 @@ public class Main {
 
     public static void aufgabe1()
     {
-        for(double i=0; i < 200; i++) {
+        for(int i=0; i < 200; i++) {
             System.out.println(i);
         }
     }
